Fix missing Usability Post link
Experiment and find out that specifying a feed both in subscriptions.xml
and in config.ini does the right thing

Discover that feed titles are supposed to be in the `name` attribute in
the INI file, and that Planet can figure out titles form the feed all by
itself.

Incidentally, when you try to override a blog's link, be sure to do

  rm cache/sources/mangledurl
  bin/planet -n

because if you just do bin/planet -o, changes will not take effect.